Revert LOWMEMKILL short to int

+++ drivers/staging/android/lowmemorykiller.c
@@ -56,7 +56,7 @@
 
 static uint32_t lowmem_debug_level = 1;
 static uint32_t lowmem_auto_oom = 1;
-static short lowmem_adj[6] = {
+static int lowmem_adj[6] = {
 	0,
 	1,
 	6,
